feat: convert components to ts and functional components
