+++ src/components/Root.tsx
@@ -6,7 +6,7 @@ import 'react-toastify/dist/ReactToastify.css';
 import i18nConfig from '../config/i18n';
 import App from './App';
 
-const Root = () => (
+const Root = (): JSX.Element => (
   <ThemeProvider theme={theme}>
     <I18nextProvider i18n={i18nConfig}>
       <App />
