+++ src/components/layout/Header.tsx
@@ -1,11 +1,12 @@
-import PropTypes from 'prop-types';
 import AppBar from '@mui/material/AppBar';
 import Typography from '@mui/material/Typography';
 import Toolbar from '@mui/material/Toolbar';
-import { withTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 import { ReactComponent as Logo } from '../../resources/logo.svg';
 
-const Header = ({ t }) => {
+const Header = (): JSX.Element => {
+  const { t } = useTranslation();
+
   return (
     <header>
       <AppBar position="static">
@@ -19,9 +20,5 @@ const Header = ({ t }) => {
     </header>
   );
 };
-Header.propTypes = {
-  t: PropTypes.func.isRequired,
-};
-const TranslatedComponent = withTranslation()(Header);
 
-export default TranslatedComponent;
+export default Header;
